+++ changelog.md
@@ -6,6 +6,7 @@
 * Add ability to use ESP8266 through SPI port.
 * Change board types on the fly!
 * Add loop function for internal timing operations
+* Add function to turn time stamps on `useTimeStamp(true)`, time stamps are disabled by default.
 
 ### Enhancements
 
@@ -13,11 +14,11 @@
 
 ### Breaking Changes
 
-* Removed boolean `useAccel` in support of new mode paradigm. To use the accel simply set `curAccelMode` to `ACCEL_MODE_ON` or a function call `useAccel()`. You can also turn off the excel with a call to `useAccel(false)`.
+* Removed boolean `useAccel` in support of new mode paradigm. To use the accel, do nothing, it's enabled by default, to not use the accel simply call `useAccel(false)` and send the contents of `auxData` array each packet.
 * Removed `OPENBCI_EOP_STND_ACCEL`, `OPENBCI_EOP_STND_RAW_AUX`, `OPENBCI_EOP_USER_DEFINED`, `OPENBCI_EOP_ACCEL_TIME_SET`, `OPENBCI_EOP_ACCEL_TIME_SYNCED`, `OPENBCI_EOP_RAW_AUX_TIME_SET`, and `OPENBCI_EOP_RAW_AUX_TIME_SYNCED` in place of `enum` support replacement: `PACKET_TYPE_ACCEL`, `PACKET_TYPE_RAW_AUX`, `PACKET_TYPE_USER_DEFINED`, `PACKET_TYPE_ACCEL_TIME_SET`, `PACKET_TYPE_ACCEL_TIME_SYNC`, `PACKET_TYPE_RAW_AUX_TIME_SET`, and `PACKET_TYPE_RAW_AUX_TIME_SYNC`.
 * Removed `sniffMode` in place for `curBoardMode == OPENBCI_BOARD_MODE_DEBUG`
-* Removed `sendChannelData(void)` function, use `sendChannelData(PACKET_TYPE)` and specify packet type.
 * Removed public `waitForNewChannelData()`
+* Removed public `timeSynced` and private `sendTimeSyncUpPacket`
 
 # v2.0.1